refactor: remove old types and duplicate code
