+++ src/components/questionary/DefaultReviewWizardStep.ts
@@ -5,7 +5,7 @@ import {
 import { StepType } from 'models/questionary/StepType';
 
 export class DefaultReviewWizardStep implements WizardStep {
-  public type: StepType = 'VisitReview';
+  public type: StepType = 'ReviewStep';
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   public payload?: any;
 
@@ -16,17 +16,14 @@ export class DefaultReviewWizardStep implements WizardStep {
   ) {}
 
   getMetadata(state: QuestionarySubmissionState) {
-    const lastProposalStep =
+    const lastStep =
       state.questionary.steps[state.questionary.steps.length - 1];
-    const hasLastProposalStep = lastProposalStep !== undefined;
-
-    const isReadOnly =
-      hasLastProposalStep && lastProposalStep.isCompleted === false;
+    const isLastStepCompleted = lastStep.isCompleted;
 
     return {
       title: 'Review',
       isCompleted: this.isReviewStepCompleted(state),
-      isReadonly: isReadOnly,
+      isReadonly: isLastStepCompleted === false,
     };
   }
 }
