+++ src/components/questionary/questionaries/proposal/ProposalQuestionaryDefinition.ts
@@ -1,5 +1,5 @@
 import ProposalSummary from 'components/proposal/ProposalSummary';
-import { DefaultReviewWizardStep } from 'components/questionary/createDefaultReviewWizardStep';
+import { DefaultReviewWizardStep } from 'components/questionary/DefaultReviewWizardStep';
 import { DefaultStepDisplayElementFactory } from 'components/questionary/DefaultStepDisplayElementFactory';
 import { DefaultWizardStepFactory } from 'components/questionary/DefaultWizardStepFactory';
 import { Sdk, TemplateGroupId } from 'generated/sdk';
